ai_engine/vibe_coding/requirements.txt
# Task Management API — Production Dependencies
# All versions pinned for reproducible builds.
# Update via: pip-compile requirements.in

# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.110.0
uvicorn[standard]==0.29.0
starlette==0.36.3

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy[asyncio]==2.0.29
asyncpg==0.29.0
alembic==1.13.1

# ── Validation & Settings ─────────────────────────────────────────────────────
pydantic==2.6.4
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Security ──────────────────────────────────────────────────────────────────
bcrypt==4.1.2
PyJWT==2.8.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
aiosqlite==0.20.0

# ── Observability ─────────────────────────────────────────────────────────────
# prometheus-fastapi-instrumentator==6.1.0  # Uncomment to enable /metrics
